finish translating all bash scripts to js
