+++ clean.js
@@ -2,7 +2,7 @@
  * This script removes intermediate and final build files
  */
 
-const fs = require("fs").promises;
+const buildTools = require("./build-tools.js");
 
 const toRemove = [
   "build",
@@ -10,10 +10,5 @@ const toRemove = [
   "release-builds"
 ];
 
-console.log("Removing the following directories: " +
-  toRemove.reduce((prev, cur) => `${prev}, ${cur}`));
+buildTools.cleanDirs(toRemove);
 
-// this fold removes each directory in a row
-toRemove.reduce((prev, cur) => {
-  return prev.then(() => fs.rmdir(cur, { recursive: true }));
-}, Promise.resolve());
